requirements.txt
flask
pyserial
flask_sqlalchemy
schedule
MySQL-python
